# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoneyKeeper API",
    "description": "This is a simple MoneyKeeper software.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://alnat.dev"
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/account/search": {
      "get": {
        "tags": [
          "Account API"
        ],
        "summary": "Получение счета по идентификатору",
        "operationId": "getAccountByKey",
        "parameters": [
          {
            "name": "accountKey",
            "in": "query",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "card"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Счет с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/account/custom": {
      "post": {
        "tags": [
          "Account API"
        ],
        "summary": "Создание нового счета по параметрам",
        "operationId": "addAccount",
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Идентификатор счета",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "test_account"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Имя счета",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "Тестовый счет"
          },
          {
            "name": "description",
            "in": "query",
            "description": "Описание счета",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Тестовый счет для тестирования"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Тип счета",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "CASH",
                "CARD"
              ]
            },
            "example": "CASH"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Счет успешно создан"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/account/": {
      "get": {
        "tags": [
          "Account API"
        ],
        "summary": "Получение списка счетов",
        "operationId": "getAccountList",
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "put": {
        "tags": [
          "Account API"
        ],
        "summary": "Обновление информации по счету",
        "operationId": "updateAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Счет успешно обновлен"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "Account API"
        ],
        "summary": "Создание нового счета",
        "operationId": "addAccount_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Account"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Счет успешно создан"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/account/{accountID}": {
      "get": {
        "tags": [
          "Account API"
        ],
        "summary": "Получение счета по идентификатору",
        "operationId": "getAccountByID",
        "parameters": [
          {
            "name": "accountID",
            "in": "path",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Счет с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Account"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "delete": {
        "tags": [
          "Account API"
        ],
        "summary": "Удаление счета",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "name": "accountID",
            "in": "path",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Счета с таким идентификатором не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Удаление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/account/info/{accountID}": {
      "get": {
        "tags": [
          "Account API"
        ],
        "summary": "Получение агрегированной информации по счету",
        "operationId": "getAccountInfo",
        "parameters": [
          {
            "name": "accountID",
            "in": "path",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          },
          {
            "name": "from",
            "in": "query",
            "description": "Дата начала выборки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2020-01-01"
          },
          {
            "name": "to",
            "in": "query",
            "description": "Дата завершения выборки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2020-01-02"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Счет с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/account/info": {
      "get": {
        "tags": [
          "Account API"
        ],
        "summary": "Получение агрегированной информации по счету",
        "operationId": "getAccountInfo_1",
        "parameters": [
          {
            "name": "accountKey",
            "in": "query",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": 1
          },
          {
            "name": "from",
            "in": "query",
            "description": "Дата начала выборки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2020-01-01"
          },
          {
            "name": "to",
            "in": "query",
            "description": "Дата завершения выборки",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2020-01-02"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Счет с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfo"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
...
    "/api/transaction/account/": {
      "get": {
        "tags": [
          "Transaction API"
        ],
        "summary": "Получение списка транзакции по счету",
        "operationId": "getTransactionList_2",
        "parameters": [
          {
            "name": "accountKey",
            "in": "query",
            "description": "Идентификатор счета",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "card"
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки"
          }
        }
      },
      "Account": {
        "required": [
          "key",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "description": "Дата создания записи",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "description": "Дата обновления записи",
            "format": "date-time"
          },
          "accountID": {
            "type": "integer",
            "description": "Идентификатор счета в БД",
            "format": "int32"
          },
          "key": {
            "type": "string",
            "description": "Идентификатор счета"
          },
          "name": {
            "type": "string",
            "description": "Имя счета"
          },
          "description": {
            "type": "string",
            "description": "Описание"
          },
          "balance": {
            "type": "number",
            "description": "Текущий баланс счета"
          },
          "type": {
            "type": "string",
            "description": "Тип счета",
            "enum": [
              "CASH",
              "CARD"
            ]
          },
          "transactionList": {
            "type": "array",
            "description": "Список проводок по счету",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        },
        "description": "Измененный счет"
      },
...
      "Transaction": {
        "required": [
          "account",
          "amount",
          "processDate",
          "status",
          "type"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "description": "Дата создания записи",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "description": "Дата обновления записи",
            "format": "date-time"
          },
          "transactionID": {
            "type": "integer",
            "format": "int32"
          },
          "processDate": {
            "pattern": "yyyy-MM-dd HH:mm:ss",
            "type": "string",
            "description": "Дата проведения",
            "format": "date-time"
          },
          "amount": {
            "type": "number",
            "description": "Сумма"
          },
          "status": {
            "type": "string",
            "description": "Статус транзакции",
            "enum": [
              "CONFORMED",
              "ROLLBACK"
            ]
          },
          "type": {
            "type": "string",
            "description": "Тип проводки",
            "enum": [
              "ADDITION",
              "SUBTRACTION",
              "MOVEMENT"
            ]
          },
          "comment": {
            "type": "string",
            "description": "Комментарий по транзакции"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "account": {
            "$ref": "#/components/schemas/Account"
          }
        },
        "description": "Новая транзакция"
      },
      "AccountBalance": {
        "type": "object",
        "properties": {
          "date": {
            "pattern": "yyyy-MM-dd",
            "type": "string",
            "description": "Дата баланса",
            "format": "date",
            "example": "2020-01-01"
          },
          "balance": {
            "type": "number",
            "description": "Баланс"
          }
        },
        "description": "Информация о балансе счета по дню"
      },
      "AccountInfo": {
        "required": [
          "from",
          "to"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/Account"
          },
          "balance": {
            "type": "array",
            "description": "Данные о балансе счета по дню",
            "items": {
              "$ref": "#/components/schemas/AccountBalance"
            }
          },
          "from": {
            "pattern": "yyyy-MM-dd",
            "type": "string",
            "description": "Дата начала выборки",
            "format": "date",
            "example": "2020-01-01"
          },
          "to": {
            "pattern": "yyyy-MM-dd",
            "type": "string",
            "description": "Дата завершения выборки",
            "format": "date",
            "example": "2020-01-10"
          },
          "transactionList": {
            "type": "array",
            "description": "Список транзакций",
            "items": {
              "$ref": "#/components/schemas/Transaction"
            }
          }
        },
        "description": "Информация о счете в агрегированном виде"
      },
...
    }
  }
}